#10658 A minor refactor to replace the API implemantation in a interface, impl
